fix: add pluginEmitRouteFile in uni-builder
