+++ packages/builder/uni-builder/package.json
@@ -30,6 +30,7 @@
     "@babel/core": "^7.23.2",
     "@babel/preset-react": "^7.22.15",
     "@babel/types": "^7.23.0",
+    "@modern-js/utils": "workspace:*",
     "@modern-js/server": "workspace:*",
     "@modern-js/prod-server": "workspace:*",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
